"Alquiler: Cambio de entidad y finalizacion del CRUD "

+++ src/main/java/com/proyectoIntegrador/gameShare/repositorio/AlquilerRepositorio.java
@@ -3,5 +3,5 @@ package com.proyectoIntegrador.gameShare.repositorio;
 import com.proyectoIntegrador.gameShare.entidad.Alquiler;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlquilerRepositorio extends JpaRepository<Alquiler, Integer> {
+public interface AlquilerRepositorio extends JpaRepository<Alquiler, Long> {
 }
